architecture changes, improvements, remove phonemes for now

+++ utils/checks.py
@@ -1,29 +1,18 @@
 import datetime as dt
-import os
-import platform
-import requests
-import shutil
-import sys
-import torch
+import os, platform, shutil, socket, torch
 
-from dotenv import load_dotenv
-from flask import  jsonify
-from pathlib import Path
-from tqdm import tqdm
-
-from .starter import batch_check
-from .voices import get_model_info
-
-available_models = [ 'samantha' ]
+from dotenv import load_dotenv as LoadDotEnv
+from flask import jsonify as FlaskJsonify
+from typing import Tuple
 
 def env_check():
+    from .envinfo import env_api_key, env_listen_port
+
     try:
-        load_dotenv()
-        env_api_key = os.getenv('API_KEY')
-        env_port = os.getenv('LISTENING_PORT')
+        LoadDotEnv()
         if not env_api_key:
             raise ValueError('API_KEY is missing in .env file')
-        if not env_port:
+        if not env_listen_port:
             raise ValueError('PORT is missing in .env file')
         return True
     except Exception as e:
@@ -31,33 +20,31 @@ def env_check():
         raise SystemExit('Missing .env file.  Exiting...')
 
 def api_check(api_key):
+     from .envinfo import env_api_key
+
      if not api_key:
-        return jsonify({'error': 'API key is missing'}), 401
+        return FlaskJsonify({'error': 'API key is missing'}), 401
      else:
         try:
-            load_dotenv()
-            env_api_key = os.getenv('API_KEY')
+            LoadDotEnv()
             if(env_api_key != api_key):
-                return jsonify({'error': 'Invalid API key'}), 401
+                return FlaskJsonify({'error': 'Invalid API key'}), 401
             else:
                 return True
         except:
-            return jsonify({'error': 'Invalid API key'}), 401
+            return FlaskJsonify({'error': 'Invalid API key'}), 401
 
-def cuda_check():
-    if not torch.cuda.is_available():
+def cuda_check(show = True):
+    cuda = torch.cuda.is_available()
+    if show and not cuda:
         print("***WARNING***  CUDA is not available.  Falling back to CPU.")
-
-from typing import Tuple
+    return cuda
 
 def find_anaconda(operating_system: str) -> Tuple[str, str] or bool:
-    home_dir = Path.home()
-    if operating_system == "Windows":
-        possible_locations = [home_dir / "Anaconda3", home_dir / "Miniconda3"]
-        helper_file = "Scripts/activate.bat"
-    else:  # Linux or MacOS
-        possible_locations = [home_dir / "anaconda3", home_dir / "miniconda3"]
-        helper_file = "bin/activate"
+    from .vars import conda_possible_locations, conda_help_file_win, conda_help_file_linux
+
+    possible_locations = conda_possible_locations
+    helper_file = conda_help_file_win if operating_system == "Windows" else conda_help_file_linux
 
     for location in possible_locations:
         if (location / helper_file).is_file():
@@ -66,11 +53,11 @@ def find_anaconda(operating_system: str) -> Tuple[str, str] or bool:
             return str_location, str_helper_file
     raise FileNotFoundError("Anaconda not found.")
 
-
 def find_venv(operating_system: str) -> Tuple[str, str] or bool:
-    home_dir = Path.home()
-    possible_locations = [home_dir / ".venvs", home_dir / "venvs", home_dir / ".virtualenvs"]
-    helper_file = "Scripts/activate.bat" if operating_system == "Windows" else "bin/activate"
+    from .vars import venv_possible_locations, venv_help_file_win, venv_help_file_linux
+
+    possible_locations = venv_possible_locations
+    helper_file = venv_help_file_win if operating_system == "Windows" else venv_help_file_linux
 
     for location in possible_locations:
         for venv in location.glob("*"):
@@ -78,7 +65,6 @@ def find_venv(operating_system: str) -> Tuple[str, str] or bool:
                 str_venv = str(venv)
                 str_helper_file = str(venv / helper_file)
                 return str_venv, str_helper_file
-
     raise FileNotFoundError("venv not found.")
 
 def folder_check(path):
@@ -87,8 +73,10 @@ def folder_check(path):
     return path
 
 def get_wav_path():
-    file_ext = ".wav"
-    folder_name = "output_wavs"
+    from .vars import wav_ext, output_wavs_folder
+
+    file_ext = wav_ext
+    folder_name = output_wavs_folder
     now = dt.datetime.now()
     filename = now.strftime("%Y%m%d_%H%M%S") + file_ext
     folder_path = os.path.join(os.getcwd(), folder_name)
@@ -108,9 +96,11 @@ def get_os():
         return None
 
 def sub_check():
+    from .vars import models_folder, output_wavs_folder
+
     try:
-        models = os.path.join(os.getcwd(), "models")
-        output_wavs = os.path.join(os.getcwd(), "output_wavs")
+        models = os.path.join(os.getcwd(), models_folder)
+        output_wavs = os.path.join(os.getcwd(), output_wavs_folder)
 
         folders = [models, output_wavs]
 
@@ -123,7 +113,10 @@ def sub_check():
         raise SystemExit('Error creating folders.  Exiting...')
     
 def batch_file_exists():
-    batch_file = os.path.join(os.getcwd(), "start.bat") or os.path.join(os.getcwd(), "start.sh")
+    from .starter import batch_check
+    from .vars import bat_ext, sh_ext, start_file
+
+    batch_file = os.path.join(os.getcwd(), (start_file + bat_ext)) or os.path.join(os.getcwd(), start_file + sh_ext)
     if os.path.exists(batch_file):
         try:
             batch_check(get_os())
@@ -135,14 +128,17 @@ def batch_file_exists():
         return False
     
 def models_check(meta_file_name, output_folder):
+    from .envinfo import env_default_model, env_custom_models
+    from .vars import models_folder, available_models
+    from .voices import get_model_info
 
     operating_system = get_os()
 
-    load_dotenv()
+    LoadDotEnv()
 
     models = []
-    default_model = os.getenv("DEFAULT_MODEL")
-    custom_models = os.getenv("CUSTOM_MODELS")
+    default_model = env_default_model
+    custom_models = env_custom_models
 
     if default_model:
         models.append(default_model)
@@ -163,7 +159,7 @@ def models_check(meta_file_name, output_folder):
             retrieve_models.append(model)
 
         for model in retrieve_models:
-            modelroot = os.getcwd() + '/' + "models"
+            modelroot = os.getcwd() + '/' + models_folder
             modelinfo = get_model_info(model)
             modelfolder = modelinfo[0]
             url = modelinfo[1]
@@ -176,39 +172,40 @@ def models_check(meta_file_name, output_folder):
             if not os.path.exists(foldername):
                 raise SystemError(f'Error with {foldername} folder.  Exiting...')
             elif os.path.exists(modelroot + '/' + file_name):
-                try:
-                    print(f'Extracting {file_name} to {foldername}...')
-                    if operating_system == 'Windows':
-                        os.system(f'powershell -Command "& \'{path_7z}\' x {modelroot}/{file_name} -o{modelroot}"')
-                    else:
-                        os.system(f"7z x {modelroot}/{file_name} -o{modelroot}")
-                    print(f'Extraction of {file_name} to {foldername} complete...')
-                except Exception as e:
-                    raise SystemExit(f'Error with {model} model and {file_name}. Exiting...')
+                if not os.path.exists(foldername + '/' + meta_file_name) and os.path.exists(foldername + '/' + output_folder):
+                    try:
+                        print(f'Extracting {file_name} to {foldername}...')
+                        if operating_system == 'Windows':
+                            os.system(f'powershell -Command "& \'{path_7z}\' x {modelroot}/{file_name} -o{modelroot}"')
+                        else:
+                            os.system(f"7z x {modelroot}/{file_name} -o{modelroot}")
+                        print(f'Extraction of {file_name} to {foldername} complete...')
+                    except Exception as e:
+                        raise SystemExit(f'Error with {model} model and {file_name}. Exiting...')
             else:
-                try:
-
-                    if operating_system == 'Windows':
-                        print(f'Downloading {file_name} to {modelroot} via powershell...')
-                        os.system(f'powershell -Command "(New-Object System.Net.WebClient).DownloadFile(\'{url}\', \'{modelroot}/{file_name}\')"')
-                        print(f'Download of {file_name} complete.')
-                    elif operating_system == 'MacOS':
-                        print(f'Downloading {file_name} to {modelroot} via curl...')
-                        os.system(f'curl -o {modelroot}/{file_name} {url}')
-                        print(f'Download of {file_name} complete.')
-                    else:
-                        print(f'Downloading {file_name} to {modelroot} via wget...')
-                        os.system(f'wget {url} -P {modelroot}')
-                        print(f'Download of {file_name} complete.')
-                    print(f'Extracting {file_name} to {foldername}...')
-                    if operating_system == 'Windows':
-                        os.system(f'powershell -Command "& \'{path_7z}\' x {modelroot}/{file_name} -o{modelroot}"')
-                    else:
-                        os.system(f"7z x {modelroot}/{file_name} -o{modelroot}")
-                    print(f'Extraction of {file_name} to {foldername} complete...')
-                except Exception as e:
-                    print(f'Error with {model} model and downloading {file_name}. Exiting...')
-                    raise SystemExit(f'Error: {e}')
+                if not os.path.exists(foldername + '/' + meta_file_name) and os.path.exists(foldername + '/' + output_folder):
+                    try:
+                        if operating_system == 'Windows':
+                            print(f'Downloading {file_name} to {modelroot} via powershell...')
+                            os.system(f'powershell -Command "(New-Object System.Net.WebClient).DownloadFile(\'{url}\', \'{modelroot}/{file_name}\')"')
+                            print(f'Download of {file_name} complete.')
+                        elif operating_system == 'MacOS':
+                            print(f'Downloading {file_name} to {modelroot} via curl...')
+                            os.system(f'curl -o {modelroot}/{file_name} {url}')
+                            print(f'Download of {file_name} complete.')
+                        else:
+                            print(f'Downloading {file_name} to {modelroot} via wget...')
+                            os.system(f'wget {url} -P {modelroot}')
+                            print(f'Download of {file_name} complete.')
+                        print(f'Extracting {file_name} to {foldername}...')
+                        if operating_system == 'Windows':
+                            os.system(f'powershell -Command "& \'{path_7z}\' x {modelroot}/{file_name} -o{modelroot}"')
+                        else:
+                            os.system(f"7z x {modelroot}/{file_name} -o{modelroot}")
+                        print(f'Extraction of {file_name} to {foldername} complete...')
+                    except Exception as e:
+                        print(f'Error with {model} model and downloading {file_name}. Exiting...')
+                        raise SystemExit(f'Error: {e}')
             if os.path.exists(foldername + '/' + meta_file_name) and os.path.exists(foldername + '/' + output_folder):
                 continue
             else:
@@ -216,13 +213,16 @@ def models_check(meta_file_name, output_folder):
     
     select_all_models_7z = os.path.join(os.getcwd(), "models", "*.7z")
 
-    if operating_system == 'Windows':
-        os.system(f'del {select_all_models_7z}')
-    elif operating_system == 'MacOS' or 'Linux':
-        os.system(f'rm {select_all_models_7z}')
+    if os.path.exists(select_all_models_7z):
+        if operating_system == 'Windows':
+            os.system(f'del {select_all_models_7z}')
+        elif operating_system == 'MacOS' or 'Linux':
+            os.system(f'rm {select_all_models_7z}')
 
 def check_chocolatey():
-    if not shutil.which('choco'):
+    from .vars import choco_file
+
+    if not shutil.which(choco_file):
         try:
             print("Chocolatey is not installed.  Installing Chocolatey...")
             os.system("powershell -Command \"Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))\"")
@@ -234,7 +234,9 @@ def check_chocolatey():
             raise SystemExit('Chocolatey installation failed.  Exiting...')
 
 def check_brew():
-    if not shutil.which('brew'):
+    from .vars import brew_file
+
+    if not shutil.which(brew_file):
         try:
             print("Homebrew is not installed.  Installing Homebrew...")
             os.system('/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"')
@@ -246,13 +248,15 @@ def check_brew():
             raise SystemExit('Homebrew installation failed.  Exiting...')
 
 def check_7zip(operating_system):
+    from .vars import choco_file, brew_file, sudo_apt_file
+
     the_path = get_7zip_path(operating_system)
     if(operating_system == "Windows"):
         if not os.path.exists(the_path):
             if check_chocolatey():
                 try:
                     print("7zip is not installed.  Installing 7zip...")
-                    os.system("choco install 7zip -y")
+                    os.system(f"{choco_file} install 7zip -y")
                     print("7zip installed successfully.")
                     return True
                 except Exception as e:
@@ -264,7 +268,7 @@ def check_7zip(operating_system):
                 if check_brew():
                     try:
                         print("7zip is not installed.  Installing 7zip...")
-                        os.system("brew install p7zip -y")
+                        os.system(f"{brew_file} install p7zip -y")
                         print("7zip installed successfully.")
                         return True
                     except Exception as e:
@@ -275,8 +279,8 @@ def check_7zip(operating_system):
             if not os.path.exists(the_path):
                 try:
                     print("7zip is not installed.  Installing 7zip...")
-                    os.system("sudo apt-get update")
-                    os.system("sudo apt-get install -y p7zip-full")
+                    os.system(f"{sudo_apt_file} update")
+                    os.system(f"{sudo_apt_file} install -y p7zip-full")
                     print("7zip installed successfully.")
                     return True
                 except Exception as e:
@@ -304,4 +308,7 @@ def get_7zip_path(operating_system):
         return manual_path
     else:
         raise SystemExit('7zip installation failed. Exiting...')
-
+    
+def port_check(port_number):
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        return s.connect_ex(('localhost', int(port_number))) == 0
